Merged PR 80: Improve behavior of Start Loop/End Loop option state
## Overview
Resolves [Task 5346](https://devops.soroco.com/Soroco%20Products/Storyboard/_workitems/edit/5346).

The "Loop" button appears if either a) the number of start and end loop steps is the same or b) the number of end loop steps is greater than the number of start loop steps. The "End Loop" button appears otherwise.

## Code
* Instead of using `operation-bar.service.ts` and just toggling the button every time it was used, keep track of loop start and end count in `stepcontrol.service.ts`.

* Loop counts are recalculated when either a) the service is assigned a new `sopSectionList` or b) a section is deleted. Otherwise, loop counts are incremented and decremented as loop steps are added and removed.

* `stepcontrol.service` now exposes an observable, `detectUnpairedStartLoop`, that emits every time loop counts are updated. It emits `true` if there is an unpaired start loop.

* `operations-bar.component` subscribes to this observable and updates the `isStartLoopDisplayed ` boolean based on the result.

* `operation-bar.service.ts` has been removed.

## Notes
`stepcontrol.service` is getting a little large - loop/loop validation might need to be split out into a separate service. I put the loop counters and methods here because they need to be called by other methods that are in this service, and because the initial loop counts rely on the list of sections in this service. Open to other ideas.