Merged PR 114: Bug 5488 - unable to select sprint in create/edit user story
## Overview

Resolves [Bug 5488](https://devops.soroco.com/Soroco%20Products/Storyboard/_workitems/edit/5488). The sprint dropdown now populates correctly in the create/edit user story dialog.

## Code

The contents of the `getSprint()` method were removed in a previous pull request. This meant that the variable that populated the sprints dropdown was no longer being updated and as such the sprints dropdown could not be interacted with.

I restored the contents of the `getSprint()` method and re-added `getSprint()` to the `refresh()` method, and functionality for this use case appears to be restored.

## Notes

I'm concerned about the async-safety of this entire service. There's a bunch of API calls that happen, and then they populate a variable on a service, but there's no infrastructure for notifying the UI that data is populated and ready to be used. It seems that this mostly works out because API responses are fast enough and users navigate the interface slowly enough that users are never trying to look at information that hasn't been populated yet, but I think this is a design smell overall and we should set aside time to refactor this service.