Merged PR 146: Bug 5498 - validate invite user fields in Edit Project dialog
## Overview

Resolves [Bug 5498](https://devops.soroco.com/Soroco%20Products/Storyboard/_workitems/edit/5498).

Adds validation indication to the "Invite User" dialog fields in the Edit Project dialog.

Three of the four fields are required: Email Address, First Name, and Role. If "Send Invitation" is clicked and any of the "Invite User" fields does not have all three required fields filled in, the fields highlight in red. Any fully-valid "Invite User" field will be submitted to the server.

For validation, we are only checking that the fields are not blank. Email address validation is not performed.

## Code

Approach:
1. Extract the "Invite User" dialog piece into its own component.
2. "invite-user-field.component" uses reactive forms to determine whether or not its current contents are valid.
3. Unfortunately, the "invite user" button is not tied to this form, and it would have required risky work to make it so (moving around HTML elements and adjusting CSS). When the "invite user" button is pressed, all instances of `InviteUserFieldComponent` in the form are told that they have been submitted. If they are invalid, their invalid fields will highlight.
4. When the "invite user" button is pressed, only valid `InviteUserFieldComponent` contents are sent to the server.

* `edit-project-dialog.component.html` and `edit-project-dialog.component.scss` - extract `InviteUserFieldComponent`.
* `edit-project-dialog.component.ts` - use `ViewChildren` to communicate with variable number of created `InviteUserFieldComponent`s. Only submit form contents when valid.
* `invite-user.model.ts` - remove redundancy in field names.

## Notes

I think there are quite a few more improvements to be made here. For example, `edit-project-dialog.component.ts` keeps around the `inviteList` property, which we only use now to generate the proper number of `InviteUserFieldComponent`s. It's possible that we could initialize an `InviteUserFieldComponent` directly from a reactive form, which could be generated in `EditProjectDialogComponent`, and use the list of reactive forms to replace both `ViewChildren` and `inviteList`. I do think we need to address the `button not tied to form` issue before this will work however.

The final notes I have are more UI/UX - one can add an infinite number of `InviteUserFieldComponent`s to this form, and even when submitting a valid one they do not disappear and there is no way to delete them. We should revisit that design decision.