Merged PR 149: 5787: Update data format in Steps
## Overview

Resolves [User Story 5787](https://devops.soroco.com/Soroco%20Products/Storyboard/_workitems/edit/5787).

Updated the keys in `step-condition` `data` and `step-loop`/`step-end-loop` `data` to match spec attached to above user story.

## Code

In addition to updating those keys, I also defined TypeScript interfaces for all of the `data` attributes of each step class.

## Notes

All of these attributes are snake-cased - which is what is generally received from the backend. We may want to have a larger conversation as to when it's valuable for us to translate between snake case and camel case, especially when we're using the exact object structure that's returned from the backend (Always? Never? Somewhere in between? I don't know the answer.).