Merged PR 271: New modal for create and edit userstory
#Overview
This PR contains the new design of the create/edit userstory along with its new endpoints. The approach consist of Angular material's dialog on top of the design created by Manjit. It also contains an autocomplete field which is used to search for assignees associated with the project. Once this PR is merged the previously created component for creating and editing assignees will be deleted as now both the actions can be performed using only one component.

###New additions
1. `userstory-create-edit-modal`
2. `modal-header`

###Mods
1. `button` - added a class to change the color when the button is disabled
2. added a model for assignee