Merged PR 313: Problems with sort and filter - bug - 6286
#Overview
The issue was when sort is open, clicking on filter will open filter on top of sort also the button background color of filter did not change properly. The issue has been now fixed  by using another `@Input()` variable for the button which toggles the class depending on the current state. Also the problem of sort opening when filter is open has also been handled.