Merged PR 373: Copy User Story
#Overview
This PR adds a feature of copy userstory. On click of copy create user story modal opens with pre-populated data and then after writing US# you can save it and it will copy a userstory with the sections and steps inside it.

##File Changes

- `userstory-card.component.html` Added a menu item Copy . The ellipse is disabled if we cannot add a story (or copy a story).
- `userstory-card.component.ts` Added an event emitter on click of copy
- `modaltype.enum.ts` To format modal name added an interface member.
- `userstory-create-edit-modal.component.ts`  Dialog modal with same data as edit but name and usNumber is changed and then sending the data as payload with an extra paramenter us_copy_id which contains id of the userstory which we want to copy.
- `reasoncodes.component.ts` Open the dialog on event trigger and refresh on response.

###Notes
Testing can be done once the backend is merged or you can checkout copy_user_story branch in backend.