Merged PR 469: 7978-unable to cancel video while upload in progress and clear cache of previous file selected to upload
The subscription was not working properly. So that is fixed now also @<SOROCO\brinda.m> mentioned about another issue i.e. once a user selects a video and uploads it, if he/she selects the same video a tries to upload that does not happen. The reason for this issue is the previously selected video is stored in the memory and it does not gets cleared. Hence I have added another line of code which manually clears the previous cache.