Merged PR 418: Copy Story feature
#Overview
This feature is for copy user story part 2 . On Sop page for a copied user story you will have a icon that tells whether this section is linked or not and whenever you try to do any action section or within a linked section a prompt is shown that asked if you want to keep linking or not. On yes it propagate everywhere and on No it delinked.

##File Changes

`right-panel.component.ts` - For edit and delete step a prompt is added.
`stepcontrol.service.ts` - Changing the section details based on response , Added an event emitter that is emitted everytime we delink and it is listened in the section-tile.component.ts
`section-title.component.ts` - Added a prompt for section edit , refresh sectionLink if event is thrown by stepcontrol.service.ts, And story userstoryNames in a formcontrol on init.
`confirm-model.service.ts` - Added Confirm modal that can be shown on the linked sections.

**Note -** Backend code can be checked from the backend branch **copy_user_story** and is not pushed in uat , so its preferred to merge this branch only after the backend pr is completed.

https://devops.soroco.com/Soroco%20Products/Storyboard/_workitems/edit/6956